remove redundant code related to text preprocessing
