+++ autokeras/text/text_supervised.py
@@ -27,18 +27,10 @@ from autokeras.nn.model_trainer import BERTTrainer
 from autokeras.supervised import SingleModelSupervised
 from autokeras.text.pretrained_bert.file_utils import PYTORCH_PRETRAINED_BERT_CACHE
 from autokeras.text.pretrained_bert.modeling import BertForSequenceClassification
+from autokeras.text.pretrained_bert.text_utils import convert_examples_to_features
 from autokeras.text.pretrained_bert.tokenization import BertTokenizer
 from autokeras.utils import get_device
 from torch.utils.data import TensorDataset, DataLoader, SequentialSampler
-
-
-class InputFeatures(object):
-    """A single set of features of data."""
-
-    def __init__(self, input_ids, input_mask, segment_ids):
-        self.input_ids = input_ids
-        self.input_mask = input_mask
-        self.segment_ids = segment_ids
 
 
 class TextClassifier(SingleModelSupervised, ABC):
@@ -160,38 +152,7 @@ class TextClassifier(SingleModelSupervised, ABC):
             all_input_masks: ndarray containing 1's or 0's based on if the tokens are real or padded.
             all_segment_ids: ndarray containing all 0's since it is a classification task.
         """
-        features = []
-        for (_, example) in enumerate(x):
-            tokens_a = self.tokenizer.tokenize(example)
-
-            if len(tokens_a) > self.max_seq_length - 2:
-                tokens_a = tokens_a[:(self.max_seq_length - 2)]
-
-            tokens = ["[CLS]"] + tokens_a + ["[SEP]"]
-            segment_ids = [0] * len(tokens)
-
-            input_ids = self.tokenizer.convert_tokens_to_ids(tokens)
-
-            input_mask = [1] * len(input_ids)
-
-            padding = [0] * (self.max_seq_length - len(input_ids))
-            input_ids += padding
-            input_mask += padding
-            segment_ids += padding
-
-            if len(input_ids) != self.max_seq_length or \
-                    len(input_mask) != self.max_seq_length or \
-                    len(segment_ids) != self.max_seq_length:
-                raise AssertionError()
-
-            features.append(InputFeatures(input_ids=input_ids,
-                                          input_mask=input_mask,
-                                          segment_ids=segment_ids))
-
-        all_input_ids = torch.tensor([f.input_ids for f in features], dtype=torch.long)
-        all_input_mask = torch.tensor([f.input_mask for f in features], dtype=torch.long)
-        all_segment_ids = torch.tensor([f.segment_ids for f in features], dtype=torch.long)
-        return all_input_ids, all_input_mask, all_segment_ids
+        return convert_examples_to_features(x, self.tokenizer, self.max_seq_length)
 
     def transform_y(self, y):
         pass
